refactor: remove Header component to streamline project structure
- Deleted the Header component from src/components/Header.tsx as part of the ongoing project restructuring. This change aims to simplify the codebase and improve maintainability.
